move swap to local
