+++ packages/web/server/api/local-routers/local-router.ts
@@ -1,7 +1,7 @@
-import { swapRouter } from "~/server/api/edge-routers/swap-router";
 import { createTRPCRouter } from "~/server/api/trpc";
 
 import { concentratedLiquidityRouter } from "./concentrated-liquidity-router";
+import { swapRouter } from "./swap-router";
 
 /**
  * This section includes tRPC functions that execute on the client-side.
