+++ packages/web/utils/trpc.ts
@@ -151,9 +151,6 @@ export const api = createTRPCNext<AppRouter>({
             [constructEdgeRouterKey("pools")]: makeSkipBatchLink(
               `${getBaseUrl()}${constructEdgeUrlPathname("pools")}`
             )(runtime),
-            [constructEdgeRouterKey("quoteRouter")]: makeSkipBatchLink(
-              `${getBaseUrl()}${constructEdgeUrlPathname("quoteRouter")}`
-            )(runtime),
             [constructEdgeRouterKey("assets")]: makeSkipBatchLink(
               `${getBaseUrl()}${constructEdgeUrlPathname("assets")}`
             )(runtime),
